Don't escape "\x[[:hex:]][[:hex:]]" in command line wildcard parsing.
This patch makes wildcard parsing of the config file leave
"\x[[:hex:]][[:hex:]]" escape sequences in place. This allows us to pass
in \x20 and similar without it being tokenized, so that it appears
intact to all levels of the parser, which we need for e.g. filesystem
labels in command line arguments.

Based on the following patches fedora has shipped:
* 41d002615 - Pass "\x[[:hex:]][[:hex:]]" straight through unmolested. (49 minutes ago) <Peter Jones>
* 3ebad1305 - Don't munge raw spaces when we're doing our cmdline escaping (#923374) (49 minutes ago) <Hans de Goede>

Resolves: rhbz#855849
Resolves: rhbz#923374

Signed-off-by: Peter Jones <pjones@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>