grub-install: Don't mark PPC machines as EFI when /boot/efi/ exists.
Currently when platform == GRUB_INSTALL_PLATFORM_POWERPC_IEEE1275, the
presence of /boot/efi used to determine if this is an EFI machine.  This
makes no sense at all, as it's a big-endian ieee1275 platform that's
never an EFI platform.

Since some OSes (such as Fedora) always put the grubenv at
/boot/efi/EFI/fedora/grubenv, this code causes it to always make the
wrong decision on those OSes.

This isn't the right thing at all, and it's difficult to know why you
might ever want it, especially since Mac PPC machines /don't have EFI/.
This patch removes the code that does this.  I don't know why it was
there, and the commit message merely states "Implement better
integration with Mac firmware.", so there's little hope I'll figure it
out.  As such, I have not replaced this with some other heuristic or
other behavior.

Resolves: rhbz#1484474

Signed-off-by: Peter Jones <pjones@redhat.com>