tcp: try to pick new port numbers each time...
If we start with a constant each time, and we reboot quicker than
TIME_WAIT, other TCP stacks try to stitch our new TCP flow into the old
one.  This does not have the outcome we want (and it crashed Wireshark
for me.)  This is especially exacerbated by the fact that our
FIN|ACK->ACK path does not seem to work.

We may still have relatively little entropy for the first one, but it's
better than an outright constant.

Signed-off-by: Peter Jones <pjones@redhat.com>