tcp: Don't accidentally scale the TCP window in grub_net_tcp_close()
Some systems see a different scale window on FIN|ACK as needing an ACK
instead of a FIN|ACK, which just puts us in a weird loop.  Also
wireshark gets confused and misidentifies the packet as a re-scale
request, which is a pretty good indicator that absolutely nothing else
does it this way.

Signed-off-by: Peter Jones <pjones@redhat.com>