Make grub_strtoul() "end" pointer have the right const params.
grub_strtoul() and grub_strtoull() don't make the /pointer/ to "end" be
const like normal implementations do, and as a result, at many places in
the tree, people appear to have abandoned trying to figure out C's
syntax and instead of fixing it, chosen to cast a mutable pointer to an
immutable array (i.e. const char *) to (char **) in order to get *end
out correctly, or similar hacks.  This confusion stems from C's syntax
here, which (quoting from C99) is:

 declarator:
  pointer[opt] direct-declarator
 direct-declarator:
  identifier
  ( declarator )
  direct-declarator [ type-qualifier-list[opt] assignment-expression[opt] ]
  ...
  direct-declarator [ type-qualifier-list[opt] * ]
  ...
 pointer:
  * type-qualifier-list[opt]
  * type-qualifier-list[opt] pointer
 type-qualifier-list:
  type-qualifier
  type-qualifier-list type-qualifier
 ...
 type-qualifier:
  const
  restrict
  volatile

So the examples go like:

const char foo;			// immutable object
const char *foo;		// mutable pointer to object
char * const foo;		// immutable pointer to mutable object
const char * const foo;		// immutable pointer to immutable object
const char const * const foo; 	// XXX extra const keyword in the middle
const char * const * const foo; // immutable pointer to immutable
				//   pointer to immutable object
const char ** const foo;	// immutable pointer to mutable pointer
				//   to immutable object

Making const left-associative for * and right-associative for everything
else may not have been the best choice ever, but here we are, so the
last of those is actually the behavior we want from strtoul()'s "end"
pointer.

This commit fixes the types for grub_strtoul() and grub_strtoull(), as
well as fixing all of the types in all of the callers, and removing all
the type-casting which itself /removes/ const on the inputs.

In one case, we have constructions that pass in a non-const pointer and
the object it points to needs to be cast to const because gcc got
moderately overzealous with its warnings:

  int n;
  char *ptr, *next;

  for (ptr = foo; *ptr; )
    {
      ...
-     n = grub_strtoull(ptr, &ptr, ...);
+     n = grub_strtoull(ptr, (const char **)&ptr, ...);
      ...
    }

I have tested that all of this compiles without relevant warnings with
gcc 9 using 'gcc -W -Wall -Wextra' (and that /much/ of it behaves
correctly), with the exception of grub-core/osdep/aros/hostdisk.c ,
which I have no idea how to build.

Signed-off-by: Peter Jones <pjones@redhat.com>