misc: Make grub_min() and grub_max() more resilient.
grub_min(a,b) and grub_max(a,b) use a relatively naive implementation
which leads to several problems:
- they evaluate their parameters more than once, including multiple
  executions of any function calls, pre- and post- increment and
  decrements, and generally all side-effects.
- the naive way to address this, such as employed by the MIN()/MAX()
  macros currently in reiserfs.c, is to declare temporary variables in a
  statement-expression.  This isn't resilient against nested uses,
  because MIN(a,MIN(b,c)) results in the same temporary variables being
  declared in two nested scopes, which may result in a build warning
  depending on your build options, among other issues.

This patch changes our implementation to use a statement-expression
inside a helper macro, and creates the symbols for the temporary
variables with __COUNTER__ (A GNU C cpp extension) and token pasting to
create uniquely named internal variables.

Unfortunately there is one minor side effect of this, which is that
since there's an statement-expression involved, gcc generates code - and
then optimizes it away - but that is enough that the following will
trigger -Wsign-compare:

  grub_size_t foo;
  foo = some_function();
  foo = grub_max(foo, 1);

This patch also fixes the one place in the tree I could find where we
make such a comparison, introduced in 21869baec152 ("mm: Preallocate
some space when adding new regions") last week.

Signed-off-by: Peter Jones <pjones@redhat.com>