nx: add memory attribute get/set API
For NX, we need to set the page access permission attributes for write
and execute permissions.

This patch adds two new primitives, grub_get_mem_attrs() and
grub_update_mem_attrs(), and associated constant definitions, to be used
for that purpose.

For most platforms, it adds a dummy implementation that returns
GRUB_ERR_NONE.  On EFI platforms, it adds a common helper function,
grub_efi_status_to_err(), which translates EFI error codes to grub error
codes, adds headers for the EFI Memory Attribute Protocol (still pending
standardization), and an implementation of the new NX primitives using
it.  On EFI platforms which implement DXE, it additionally falls back to
using that API if the EFI API is unavailable.  This enables support on
some machines with NX capabilities that predate the EFI feature, such as
the Windows Dev Kit 2023.

Signed-off-by: Peter Jones <pjones@redhat.com>
Signed-off-by: Robbie Harwood <rharwood@redhat.com>