Add some PE flags and helper structures
Our PE support is missing several definitions from the spec, and also
has some structures specified in a way that could be more useful.

This patch adds the full list of section flags, the flag for NX
compatibility, the virtual_size field of the section table, and some
utility structs for the PE headers.

Signed-off-by: Peter Jones <pjones@redhat.com>