x86 kernel loader: allocate loader context in grub_cmd_linux()
Later we'll need information that's best gathered in grub_cmd_linux() in
order to make decisions in grub_linux_boot().

This patch changes that context to be an allocation in grub_cmd_linux(),
and switches grub_cmd_linux() to use grub_loader_set_ex() to support
that.  Additionally, it changes grub_linux_unload() to use
grub_loader_unset() to avoid double-freeing it.

Signed-off-by: Peter Jones <pjones@redhat.com>